Re-name EventListItem, updated snapshots
