--- src/components/EventListItem/EventListItem.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function EventListItem() {
-  return (
-    <div>
-
-    </div>
-  );
-}
-
-export default EventListItem;
